Add support for component labels in stepper
