+++ src/components/safe-apps/SafeAppsTxModal/styles.module.css
@@ -1,0 +1,4 @@
+.modalLabel {
+  height: 24px;
+  margin-right: 10px;
+}
