+++ src/components/safe-apps/SafeAppsTxModal/utils.ts
@@ -1,0 +1,17 @@
+import { validateAddress } from '@/utils/validation'
+import { BaseTransaction } from '@gnosis.pm/safe-apps-sdk'
+
+const validateTransaction = (t: BaseTransaction): boolean => {
+  if (!['string', 'number'].includes(typeof t.value)) {
+    return false
+  }
+
+  if (typeof t.value === 'string' && !/^(0x)?[0-9a-f]+$/i.test(t.value)) {
+    return false
+  }
+
+  const isAddressValid = validateAddress(t.to) === undefined
+  return isAddressValid && !!t.data && typeof t.data === 'string'
+}
+
+export const isTxValid = (txs: BaseTransaction[]) => txs.length && txs.every((t) => validateTransaction(t))
